+++ commons/dimas-core/src/utils.rs
@@ -4,13 +4,9 @@
 //!
 
 // region:		--- modules
-use crate::{error::Result, prelude::DimasError};
+use crate::error::{DimasError, Result};
 use dirs::{config_dir, config_local_dir, home_dir};
-use std::{
-	env,
-	sync::{mpsc::Receiver, Mutex},
-	time::Duration,
-};
+use std::env;
 // endregion:	--- modules
 
 // region:		--- functions
@@ -21,6 +17,8 @@ use std::{
 ///  - `.config` directory below home directory
 ///  - local config directory (`Linux`: `$XDG_CONFIG_HOME` or `$HOME/.config` | `Windows`: `{FOLDERID_LocalAppData}` | `MacOS`: `$HOME/Library/Application Support`)
 ///  - config directory (`Linux`: `$XDG_CONFIG_HOME` or `$HOME/.config` | `Windows`: `{FOLDERID_RoamingAppData}` | `MacOS`: `$HOME/Library/Application Support`)
+/// # Errors
+///
 pub fn find_config_file(filename: &str) -> Result<std::path::PathBuf> {
 	// handle environment path current working directory `CWD`
 	if let Ok(cwd) = env::current_dir() {
@@ -55,40 +53,6 @@ pub fn find_config_file(filename: &str) -> Result<std::path::PathBuf> {
 	Err(DimasError::FileNotFound(filename.into()).into())
 }
 // endregion:	--- functions
-
-// region:		--- TaskSignal
-#[derive(Debug, Clone)]
-/// Internal signals, used by panic hooks to inform the [`Agent`] that someting has happened.
-pub enum TaskSignal {
-	/// Restart a certain liveliness subscriber, identified by its key expression.
-	#[cfg(feature = "liveliness")]
-	RestartLiveliness(String),
-	/// Restart a certain queryable, identified by its key expression.
-	#[cfg(feature = "queryable")]
-	RestartQueryable(String),
-	/// Restart a certain lsubscriber, identified by its key expression.
-	#[cfg(feature = "subscriber")]
-	RestartSubscriber(String),
-	/// Restart a certain timer, identified by its key expression.
-	#[cfg(feature = "timer")]
-	RestartTimer(String),
-	/// just to avoid warning messages when no feature is selected.
-	#[allow(dead_code)]
-	Dummy,
-}
-
-/// Wait non-blocking for [`TaskSignal`]s.<br>
-/// Used by the `select!` macro within the [`Agent`]s main loop in [`Agent::start`].
-pub async fn wait_for_task_signals(rx: &Mutex<Receiver<TaskSignal>>) -> Box<TaskSignal> {
-	loop {
-		if let Ok(signal) = rx.lock().expect("snh").try_recv() {
-			return Box::new(signal);
-		};
-		// TODO: maybe there is a better solution than sleep
-		tokio::time::sleep(Duration::from_millis(1)).await;
-	}
-}
-// endregion:	--- TaskSignal
 
 // region:    --- Tracing
 /// Initialize tracing