+++ CHANGELOG.md
@@ -19,8 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.2] - 2024-??-??
 
 ### Added
+- dimasctl: binary to control DiMAS entities
+- dimasmon: binary to monitor DiMAS entities
 
 ### Changed
+- moved communicator & config into separate crate 'dimas-com' 
 
 ### Fixed